requirements.txt
psutil
pygithub
requests
aiohttp
async-timeout
attrs
beautifulsoup4
chardet
crontab
idna
idna-ssl
multidict
PyYAML
websockets
yarl
discord
